+++ frontend/src/common/components/absolute-background/components/squares/index.tsx
@@ -1,0 +1,21 @@
+import React from "react";
+
+const Squares = () => {
+
+  return (
+    <>
+      <div className="squares top-left" />
+      <div className="squares top-right" />
+      <div className="squares bottom-left" />
+      <div className="squares center-bottom-left" />
+      <div className="squares center-bottom-right" />
+      <div className="squares center-top-right" />
+      <div className="squares bottom-right" />
+    </>
+  );
+}
+
+export default Squares;
+
+
+