+++ frontend/src/common/components/absolute-background/index.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import Squares from "./components/squares";
+
+const AbsoluteBackground = () => {
+
+  return (
+    <>
+      <Squares/>
+    </>
+  );
+}
+
+export default AbsoluteBackground;