+++ frontend/src/pages/landing/index.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import AbsoluteBackground from '../../common/components/absolute-background';
+import Login from './login';
+
+
+const LandingPage = () => {
+
+  return (
+    <>
+      <AbsoluteBackground />
+      <Login />
+    </>
+  );
+};
+
+export default LandingPage;